requirements.txt
asgiref==3.4.1
dj-rest-auth==2.2.6
Django==3.2.16
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
python-decouple==3.6
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.1.1
